Create the to_parsed_lang module.
