--- sway-core/src/ast_transformation/mod.rs
@@ -1,1 +1,0 @@
-
