+++ sway-core/src/language/parsed/declaration/constant.rs
@@ -1,14 +1,14 @@
 use crate::{
     language::{parsed::Expression, Visibility},
+    transform,
     type_system::TypeInfo,
-    AttributesMap,
 };
 use sway_types::{Ident, Span};
 
 #[derive(Debug, Clone)]
 pub struct ConstantDeclaration {
     pub name: Ident,
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
     pub type_ascription: TypeInfo,
     pub type_ascription_span: Option<Span>,
     pub value: Expression,
