+++ sway-core/src/language/parsed/declaration/enum.rs
@@ -1,10 +1,10 @@
-use crate::{language::Visibility, type_system::*, AttributesMap};
+use crate::{language::Visibility, transform, type_system::*};
 use sway_types::{ident::Ident, span::Span};
 
 #[derive(Debug, Clone)]
 pub struct EnumDeclaration {
     pub name: Ident,
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
     pub type_parameters: Vec<TypeParameter>,
     pub variants: Vec<EnumVariant>,
     pub(crate) span: Span,
@@ -14,7 +14,7 @@ pub struct EnumDeclaration {
 #[derive(Debug, Clone)]
 pub struct EnumVariant {
     pub name: Ident,
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
     pub type_info: TypeInfo,
     pub type_span: Span,
     pub(crate) tag: usize,
