+++ sway-core/src/language/parsed/declaration/storage.rs
@@ -1,11 +1,11 @@
-use crate::{language::parsed::Expression, type_system::*, AttributesMap};
+use crate::{language::parsed::Expression, transform, type_system::*};
 use sway_types::{ident::Ident, span::Span};
 
 #[derive(Debug, Clone)]
 /// A declaration of contract storage. Only valid within contract contexts.
 /// All values in this struct are mutable and persistent among executions of the same contract deployment.
 pub struct StorageDeclaration {
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
     pub fields: Vec<StorageField>,
     pub span: Span,
 }
@@ -17,7 +17,7 @@ pub struct StorageDeclaration {
 #[derive(Debug, Clone)]
 pub struct StorageField {
     pub name: Ident,
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
     pub type_info: TypeInfo,
     pub type_info_span: Span,
     pub initializer: Expression,
