+++ sway-core/src/language/parsed/declaration/struct.rs
@@ -1,14 +1,14 @@
 use crate::{
     language::Visibility,
+    transform,
     type_system::{TypeInfo, TypeParameter},
-    AttributesMap,
 };
 use sway_types::{ident::Ident, span::Span};
 
 #[derive(Debug, Clone)]
 pub struct StructDeclaration {
     pub name: Ident,
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
     pub fields: Vec<StructField>,
     pub type_parameters: Vec<TypeParameter>,
     pub visibility: Visibility,
@@ -18,7 +18,7 @@ pub struct StructDeclaration {
 #[derive(Debug, Clone)]
 pub struct StructField {
     pub name: Ident,
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
     pub type_info: TypeInfo,
     pub(crate) span: Span,
     pub type_span: Span,
