+++ sway-core/src/language/parsed/declaration/trait.rs
@@ -1,12 +1,12 @@
 use super::{FunctionDeclaration, FunctionParameter};
 
-use crate::{language::*, type_system::TypeInfo, AttributesMap};
+use crate::{language::*, transform, type_system::TypeInfo};
 use sway_types::{ident::Ident, span::Span, Spanned};
 
 #[derive(Debug, Clone)]
 pub struct TraitDeclaration {
     pub name: Ident,
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
     pub interface_surface: Vec<TraitFn>,
     pub methods: Vec<FunctionDeclaration>,
     pub(crate) supertraits: Vec<Supertrait>,
@@ -27,7 +27,7 @@ impl Spanned for Supertrait {
 #[derive(Debug, Clone)]
 pub struct TraitFn {
     pub name: Ident,
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
     pub purity: Purity,
     pub parameters: Vec<FunctionParameter>,
     pub return_type: TypeInfo,
