+++ sway-core/src/language/ty/declaration/constant.rs
@@ -2,7 +2,7 @@ use sway_types::Ident;
 
 use crate::{
     language::{ty::*, Visibility},
-    AttributesMap,
+    transform,
 };
 
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -10,5 +10,5 @@ pub struct TyConstantDeclaration {
     pub name: Ident,
     pub value: TyExpression,
     pub(crate) visibility: Visibility,
-    pub attributes: AttributesMap,
+    pub attributes: transform::AttributesMap,
 }
