+++ sway-core/src/transform/mod.rs
@@ -1,0 +1,4 @@
+mod attribute;
+pub(crate) mod to_parsed_lang;
+
+pub use attribute::*;
