+++ sway-core/src/transform/to_parsed_lang/mod.rs
@@ -1,0 +1,5 @@
+mod convert_parse_tree;
+mod error;
+
+pub(crate) use convert_parse_tree::*;
+pub(crate) use error::*;
