+++ sway-lsp/src/utils/attributes.rs
@@ -1,8 +1,8 @@
 #![allow(dead_code)]
 use crate::core::token::{AstToken, Token};
-use sway_core::{language::parsed::Declaration, Attribute, AttributeKind, AttributesMap};
+use sway_core::{language::parsed::Declaration, transform};
 
-pub(crate) fn attributes_map(token: &Token) -> Option<&AttributesMap> {
+pub(crate) fn attributes_map(token: &Token) -> Option<&transform::AttributesMap> {
     match &token.parsed {
         AstToken::Declaration(declaration) => match declaration {
             Declaration::EnumDeclaration(decl) => Some(&decl.attributes),
@@ -20,14 +20,14 @@ pub(crate) fn attributes_map(token: &Token) -> Option<&AttributesMap> {
     }
 }
 
-pub(crate) fn doc_attributes(token: &Token) -> Option<&[Attribute]> {
+pub(crate) fn doc_attributes(token: &Token) -> Option<&[transform::Attribute]> {
     attributes_map(token)
-        .and_then(|attributes| attributes.get(&AttributeKind::Doc))
+        .and_then(|attributes| attributes.get(&transform::AttributeKind::Doc))
         .map(Vec::as_slice)
 }
 
-pub(crate) fn storage_attributes(token: &Token) -> Option<&[Attribute]> {
+pub(crate) fn storage_attributes(token: &Token) -> Option<&[transform::Attribute]> {
     attributes_map(token)
-        .and_then(|attributes| attributes.get(&AttributeKind::Storage))
+        .and_then(|attributes| attributes.get(&transform::AttributeKind::Storage))
         .map(Vec::as_slice)
 }
